refactor(configurator): split component adapter modules
